🔨 Refactor : 함수 분리
