+++ client/src/pages/OAuthLogin/OAuthLogin.module.css
@@ -28,8 +28,26 @@
 
 .errorsMsg {
   color: red;
+  margin-right: auto;
+  margin-left: 1.4rem;
+  margin-top: -0.8rem;
+  display: flex;
+  align-items: center;
+}
+
+.errorsMsg p {
+  margin-left: 0.2rem;
 }
 
 .successMsg {
-  color: blue;
+  color: #82c814;
+  margin-right: auto;
+  margin-left: 1.4rem;
+  margin-top: -0.8rem;
+  display: flex;
+  align-items: center;
+}
+
+.successMsg p {
+  margin-left: 0.2rem;
 }
